update: delete-btn color and confirm modal

+++ pages/thanks.vue
@@ -54,7 +54,9 @@
 
       <!-- buttons -->
       <div class="buttons">
-        <div class="btn" @click="deleteBookings">DELETE ALL RESERVATION</div>
+        <div class="btn btn-danger" @click="deleteBookings">
+          DELETE ALL RESERVATION
+        </div>
         <div class="btn btn-secondary" @click="$router.push('/')">
           HOMEPAGE
         </div>
@@ -75,14 +77,14 @@ export default {
   },
   methods: {
     async deleteBookings() {
-      await this.$store.dispatch('rooms/deleteBookings').then((res) => {
-        // console.log('delete res', res)
-        if (res.data.success) {
-          alert('刪除成功！')
-          this.$router.push('/')
-        }
-      })
-      // console.log('delete res', res)
+      if (window.confirm('Do you really want to delete all bookings?')) {
+        await this.$store.dispatch('rooms/deleteBookings').then((res) => {
+          if (res.data.success) {
+            alert('刪除成功！')
+            this.$router.push('/')
+          }
+        })
+      }
     },
   },
 }
